+++ gridstatus/pjm.py
@@ -15,7 +15,7 @@ from gridstatus.decorators import (
     support_date_range,
 )
 from gridstatus.lmp_config import lmp_config
-from gridstatus.pjm_dataviewer import PJMDataViewer
+from gridstatus.pjm_dataviewer_lmp_session import DataViewerLMPSession
 
 DATAVIEWER_LMP_URL = "https://dataviewer.pjm.com/dataviewer/pages/public/lmp.jsf"
 
@@ -359,22 +359,22 @@ class PJM(ISOBase):
         verbose=False,
     ):
         """Get latest LMP data from Data Viewer, which includes RT & DA"""
-        dataviewer = PJMDataViewer(self)
-        df = dataviewer._dv_lmp_fetch_data(verbose=verbose)
-        if market in (Markets.DAY_AHEAD_HOURLY, Markets.REAL_TIME_5_MIN):
-            df = df[df["Market"] == market.value]
+        with DataViewerLMPSession(pjm=self, verbose=verbose) as dv_lmp_session:
+            df = dv_lmp_session.fetch_chart_df(verbose=verbose)
+            if market in (Markets.DAY_AHEAD_HOURLY, Markets.REAL_TIME_5_MIN):
+                df = df[df["Market"] == market.value]
 
-        if end is None:
-            df = df[df["Time"].dt.date == date.date()]
-        else:
-            df = df[
-                df["Time"].dt.date.between(
-                    date.date(),
-                    end.date(),
-                    inclusive="both",
-                )
-            ]
-        return df
+            if end is None:
+                df = df[df["Time"].dt.date == date.date()]
+            else:
+                df = df[
+                    df["Time"].dt.date.between(
+                        date.date(),
+                        end.date(),
+                        inclusive="both",
+                    )
+                ]
+            return df
 
     def _get_lmp_via_pjm_json(
         self,
